+++ content/src/repository/extensions/DenylistRepository.ts
@@ -24,7 +24,6 @@ export class DenylistRepository {
     ])
   }
 
-  // This could be too
   async getAllDenylistedTargets(): Promise<{ target: DenylistTarget; metadata: DenylistMetadata }[]> {
     const result = await this.db.any(`
             SELECT DISTINCT ON (denylist.target_type, denylist.target_id)
@@ -75,7 +74,6 @@ export class DenylistRepository {
     })
 
     // Perform the query
-    // this is!!
     const queryResult = await this.db.any(
       `SELECT target_type, target_id FROM denylist WHERE ${orClause.join(' OR ')}`,
       values
